requirements.txt
flask==3.1.1
flask-socketio==5.5.1
opencv-python-headless==4.12.0.88
face-recognition==1.3.0
pillow==11.3.0
numpy==2.3.1
dlib==19.24.6
face-recognition-models==0.3.0
python-socketio==5.13.0
python-engineio==4.12.2